fix formatMsats implicit type expectations
