+++ api/paidAction/receive.js
@@ -1,7 +1,6 @@
 import { PAID_ACTION_PAYMENT_METHODS } from '@/lib/constants'
-import { toPositiveBigInt } from '@/lib/validate'
+import { toPositiveBigInt, numWithUnits, msatsToSats, satsToMsats } from '@/lib/format'
 import { notifyDeposit } from '@/lib/webPush'
-import { numWithUnits, msatsToSats, satsToMsats } from '@/lib/format'
 import { getInvoiceableWallets } from '@/wallets/server'
 import { assertBelowBalanceLimit } from './lib/assert'
