+++ lib/validate.js
@@ -513,79 +513,3 @@ export const deviceSyncSchema = object().shape({
       return true
     })
 })
-
-// check if something is _really_ a number.
-// returns true for every number in this range: [-Infinity, ..., 0, ..., Infinity]
-export const isNumber = x => typeof x === 'number' && !Number.isNaN(x)
-
-/**
- *
- * @param {any | bigint} x
- * @param {number} min
- * @param {number} max
- * @returns {number}
- */
-export const toNumber = (x, min = Number.MIN_SAFE_INTEGER, max = Number.MAX_SAFE_INTEGER) => {
-  if (typeof x === 'undefined') {
-    throw new Error('value is required')
-  }
-  if (typeof x === 'bigint') {
-    if (x < BigInt(min) || x > BigInt(max)) {
-      throw new Error(`value ${x} must be between ${min} and ${max}`)
-    }
-    return Number(x)
-  } else {
-    const n = Number(x)
-    if (isNumber(n)) {
-      if (x < min || x > max) {
-        throw new Error(`value ${x} must be between ${min} and ${max}`)
-      }
-      return n
-    }
-  }
-  throw new Error(`value ${x} is not a number`)
-}
-
-/**
- * @param {any | bigint} x
- * @returns {number}
- */
-export const toPositiveNumber = (x) => toNumber(x, 0)
-
-/**
- * @param {any} x
- * @param {bigint | number} [min]
- * @param {bigint | number} [max]
- * @returns {bigint}
- */
-export const toBigInt = (x, min, max) => {
-  if (typeof x === 'undefined') throw new Error('value is required')
-
-  const n = BigInt(x)
-  if (min !== undefined && n < BigInt(min)) {
-    throw new Error(`value ${x} must be at least ${min}`)
-  }
-
-  if (max !== undefined && n > BigInt(max)) {
-    throw new Error(`value ${x} must be at most ${max}`)
-  }
-
-  return n
-}
-
-/**
- * @param {number|bigint} x
- * @returns {bigint}
- */
-export const toPositiveBigInt = (x) => {
-  return toBigInt(x, 0)
-}
-
-/**
- * @param {number|bigint} x
- * @returns {number|bigint}
- */
-export const toPositive = (x) => {
-  if (typeof x === 'bigint') return toPositiveBigInt(x)
-  return toPositiveNumber(x)
-}
